+++ src/models/lstm.py
@@ -3,18 +3,11 @@ from time import time
 import tensorflow as tf
 from tensorflow.keras.preprocessing.text import Tokenizer
 from tensorflow.keras.preprocessing.sequence import pad_sequences
-import pandas as pd
 
 from src.utils.save_history import save_history
 from src.utils.export_model import export_model
 from src.utils.read_params import read_params
-
-
-def custom_read_csv(file_path):
-    df = pd.read_csv(file_path)
-    sentences = df.iloc[:, 0].to_numpy()
-    labels = df.iloc[:, 1].astype('category').cat.codes.to_numpy()
-    return sentences, labels
+from src.utils.custom_read_csv import custom_read_csv
 
 
 def get_tokenizer(training_sentences, vocab_size, oov_tok):
